+++ core-bundle/src/Resources/contao/models/ContentModel.php
@@ -385,7 +385,7 @@ class ContentModel extends \Model
 			$arrColumns = array("$t.pid=? AND $t.ptable=?");
 		}
 
-		if (!BE_USER_LOGGED_IN)
+		if (isset($arrOptions['ignoreFePreview']) || !BE_USER_LOGGED_IN)
 		{
 			$time = \Date::floorToMinute();
 			$arrColumns[] = "($t.start='' OR $t.start<='$time') AND ($t.stop='' OR $t.stop>'" . ($time + 60) . "') AND $t.invisible=''";
@@ -423,7 +423,7 @@ class ContentModel extends \Model
 			$arrColumns = array("$t.pid=? AND $t.ptable=?");
 		}
 
-		if (!BE_USER_LOGGED_IN)
+		if (isset($arrOptions['ignoreFePreview']) || !BE_USER_LOGGED_IN)
 		{
 			$time = \Date::floorToMinute();
 			$arrColumns[] = "($t.start='' OR $t.start<='$time') AND ($t.stop='' OR $t.stop>'" . ($time + 60) . "') AND $t.invisible=''";