+++ core-bundle/CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ### 4.2.0-beta1 (2016-XX-XX)
 
+ * Add the "ignoreFePreview" flag to the model classes (see #452).
  * Clean up the code of the file selector (see #456).